Read functions finished, update done

+++ SalesApplication/SalesApplication/ProductDetails.cs
@@ -13,7 +13,7 @@ namespace SalesApplication
         public long ID { get; set; }
         public string Name { get; set; }
         public int SaleQuantity { get; set; }
-        public float IndividualItemPrice { get; set; }
+        public decimal IndividualItemPrice { get; set; }
         public DateTime datetime { get; set; }
 
         public override string ToString()
